pipeline/pyproject.toml
[project]
name = "crawler-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "crawl4ai>=0.5.0.post8",
    "fastapi[all,standard]>=0.115.12",
    "freezegun>=1.5.1",
    "inquirer>=3.4.0",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "playwright>=1.51.0",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
    "python-dotenv>=1.1.0",
    "responses>=0.25.7",
    "rich>=14.0.0",
    "sse-starlette>=2.3.4",
    "uvicorn[standard]>=0.34.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
log_cli = true
log_cli_level = "INFO"
